requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.2.0
aiohttp>=3.8.0
yaml>=6.0
tenacity>=8.2.0
python-dotenv>=0.21.0
requests>=2.28.0
pyyaml>=6.0
httpx>=0.23.0
ujson>=5.4.0
loguru>=0.6.0